+++ packages/lib/zupoll-shared/src/makePodboxGroupUrl.ts
@@ -1,0 +1,14 @@
+import urljoin from "url-join";
+
+export function makePodboxGroupUrl(
+  ZUPASS_SERVER_URL: string,
+  pipelineId: string,
+  groupId: string
+): string {
+  return urljoin(
+    ZUPASS_SERVER_URL,
+    "generic-issuance/api/semaphore/",
+    pipelineId,
+    groupId
+  );
+}
